pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "screw-detector"
version = "0.1.0"
description = "YOLOv8 + SAHI Detection Pipeline for Tiny Object Optimization"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Screw Detector Team", email = "contact@screwdetector.dev"}
]
keywords = [
    "object-detection",
    "yolo",
    "sahi",
    "computer-vision",
    "deep-learning",
    "tiny-objects",
    "screw-detection"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
]

dependencies = [
    "ultralytics>=8.4.0",
    "sahi>=0.11.0",
    "opencv-python>=4.8.0",
    "pandas>=2.0.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0.0",
    "matplotlib>=3.7.0",
    "scipy>=1.10.0",
    "seaborn>=0.12.0",
    "psutil>=5.9.0",
    "onnx>=1.14.0",
    "onnxslim>=0.1.0",
    "onnxruntime>=1.15.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "shapely>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "black>=23.7.0",
    "mypy>=1.5.0",
    "pre-commit>=3.5.0",
]
notebook = [
    "jupyterlab>=4.0.0",
    "ipykernel>=6.25.0",
]

[project.urls]
Homepage = "https://github.com/3bsalam-1/screw-detector"
Documentation = "https://github.com/3bsalam-1/screw-detector#readme"
Repository = "https://github.com/3bsalam-1/screw-detector"
"Bug Tracker" = "https://github.com/3bsalam-1/screw-detector/issues"

[project.scripts]
screw-train = "src.scripts.train:main"
screw-evaluate = "src.scripts.evaluate:main"
screw-export = "src.scripts.export:main"
screw-demo = "src.scripts.demo:main"
screw-slice-dataset = "src.scripts.slice_dataset:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"screw_detector" = ["py.typed"]

# Ruff configuration
[tool.ruff]
target-version = "py39"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

# Black configuration
[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# MyPy configuration
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "sahi.*",
    "ultralytics.*",
    "cv2.*",
    "yaml.*",
]
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

# Coverage configuration
[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
